requirements.txt
numpy==1.23.1
opencv-python==4.6.0.66
pika==1.3.0
PyYAML==6.0
